Fixed issue 220: Few issues with static code analysis
git-svn-id: http://crashrpt.googlecode.com/svn/trunk@1632 821a4644-a21a-0410-aec7-6199d47568cc

+++ processing/crashrptprobe/MinidumpReader.cpp
@@ -771,9 +771,9 @@ int CMiniDumpReader::StackWalk(DWORD dwThreadId)
 
         if(!sSourceFile.IsEmpty())
         {
-            size_t pos = sSourceFile.ReverseFind('\\');
+            int pos = sSourceFile.ReverseFind('\\');
             if(pos>=0)
-                sSourceFile = sSourceFile.Mid((int)pos+1);
+                sSourceFile = sSourceFile.Mid(pos+1);
             str += _T(" [ ");
             str += sSourceFile;
             str += _T(": ");
